Apply all filters from current record list when exporting, and adjust Export UI/UX
