+++ locale/en/corteza-webapp-compose/block.yaml
@@ -295,12 +295,12 @@ recordList:
     all: Export all records
     allow: Users will be able to export records
     csv: CSV Export
-    dateRange: 'Select date range:'
+    dateRange: Select date range
     query:
       label: Search query
       placeholder: Search...
     filter:
-      label: Filter records
+      label: Record filter
       placeholder: field1 = 1 AND field2 = 232
       footnote: Simplified SQL condition (WHERE ...) syntax is supported. Variables like {{0}}, {{1}} and {{2}} are evaluated (when available)
       createdAt: Record created
@@ -310,12 +310,14 @@ recordList:
       thisMonth: This month
       thisWeek: This week
       today: Today
+      from: From
+      to: To
       updatedAt: Record updated
-    inRange: Set date range
+    inRange: Date range
     includeQuery: Filter by search query
     json: JSON Export
     limitations: 'CSV export limitation: only the first value in the multi value fields will be exported'
-    rangeBy: 'Set range by:'
+    rangeBy: Set range by
     recordCount: '{{count}} records ready for export'
     selectFields: 'Select fields you want to export:'
     selection: Selected records
